backend/requirements.txt
faiss-cpu
sentence-transformers
google-generativeai
fastapi
uvicorn
pydantic
